requirements.txt
Flask_Security
passlib
PTable
Cython
Flask_Mail
numpy
requests
Werkzeug
SQLAlchemy
Jinja2
Flask
Flask_Login
Flask_SQLAlchemy
BeneDict
flask_admin
flask_bootstrap
flask_cors
flask_restful
prettytable
sqlalchemy_utils
kubernetes
psycopg2-binary
prometheus_client
ansi2html
colorama
psutil
tqdm
pytest-order
onnx
onnxruntime
keras2onnx
atlas-mpl-style
scikit-learn
seaborn
jupyterlab
wandb
pandarallel
tables
colorlog
pybeamer
rootplotlib
tf2onnx

